Fix VC8 compilation
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@8028 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/StdAfx.cpp
@@ -71,7 +71,7 @@ CString& CLowerCaseTable::operator()(CString& strSource) const
 		regex::subst_results results;
 		std::wstring strTemp( strSource, strSource.GetLength() );
 		regExpPattern.substitute( strTemp, results );
-		strSource.SetString( strTemp.c_str(), strTemp.length() );
+		strSource.SetString( strTemp.c_str(), (int)strTemp.length() );
 	}
 
 	// Lowercase now everything. Not final sigmas are taken from the table
